drivers: rmnet_shs: Tune for  UDP GRO and logging
UDP GRO causes packet sizes to grow. Packet thresholds need to be
adjusted to correct for this and allow for optimal switching.

Add a module param to gate trace debugging. Causes traces not to
be peeked unless module param is set.

Pick cores based on most available core as opposed to an active core
that still has room as load will grow and can lead to overloading.

CRs-Fixed: 2361081
Change-Id: Icad2a1f3eef82ea2134747c0cb7115273254e697
Acked-by: Raul Martinez <mraul@qti.qualcomm.com>
Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>