drivers: perf: Record number of packets dropped due to OOM
Currently, if allocating a large SKB for the flow fails, we simply
print an error to the dmesg log. This doesn't give us any useful
information, and can cause log spam in low-memory situations. Instead
of printing this message, track the number of packets we've dropped
because of the allocation failure.

Change-Id: I0724dcfb9d670eada17d9623dfa5fe1279e6a0de
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>