drivers: shs: Fix spinlock recursion during flushing
If flushing WQ gets prempted on Core 1 after it locks hashtable spinlock
it is possible for NET_RX irq to occur on Core 1 and try to relock the
hashtable spinlock again.

This fix disables bottom half when we are flushing in wq context preventing
NET_RX from occuring on the same core. This shouldn't happen unless pkts
are parked longer than the limit and wq is created to flush.

CRs-Fixed: 2365477
Change-Id: I3b3000ff3f1d75c934917c2ed1e3aafcb7bacdd4
Acked-by: Raul Martinez <mraul@qti.qualcomm.com>
Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>