drivers: perf: Reserve headroom on allocated SKBs
SKBs allocated by the rmnet_perf module have extra space allocated for
additional headers to be prepended or appended by the stack during
processing. However, the necessary headroom was never reserved before
filling the SKB data, so attempting to add headers to the beginning of the
packet (such as Ethernet) could fail.

Change-Id: I4af046aa277c78eaca139f2a3cec67cc0669c8e0
CRs-Fixed: 2366759
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>