drivers: rmnet_shs: DL ind failsafe, timer tweak
Previously early packets heavily relied on timer failsafe
to be delivered correctly as DL ind is disabled. This could
cause OOO packets if a packet without a dl trailer is being
flushed by wq and subsequent packets with dl trailer come
during the flush. The gro_cells IRQ from the wq flush would
be racing with the NET_RX IRQ potentially causing OOO packets.

Change adds a failsafe for packets that arrive without dl headr
so that the timer failsafe is not needed. It also increases timer
duration so it triggers only on the most serious of delays.

CRs-Fixed: 2371786
Change-Id: I64df4d7fdea527fe917e154f3492dd211d092fec
Acked-by: Raul Martinez <mraul@qti.qualcomm.com>
Signed-off-by: Conner Huff <chuff@codeaurora.org>