drivers: rmnet_perf: flush UDP flows on OOO IP packets
In order to ensure correct header reconstruction of coalesced packet
frames, rmnet_perf must end coalescing when detecting packets are out of
order. Since UDP headers have no ordering information, we need to use the
ID field of the IPv4 header to detect this, and flush out the flows when
the packet we're processing is found to be out of sequence.

Processing for UDP flows using IPv6 remains unchanged, as IPv6 contains
no fields indicative of ordering that we could leverage for this purpose.

Change-Id: Id52e8138f30e1812a94445fd45981dc2c5bc772d
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>