drivers: rmnet_shs: Cancel flush wq during cleanup
Previously we did not clean up any packet flushing work queued
on the global workqueue. This could lead to a issue if there is
a core switch work or a failsafe flush work queued that manages
to run after the module deinit as the memory was deallocated.

This change prevents that by cancelling any queued work on the
global workqueue before final deinitialization.

Rename work structs with more accurate names.

CRs-Fixed: 2382173
Change-Id: I4ffe1842a1852d0e3f60dda987198b1571cbfee6
Acked-by: Raul Martinez <mraul@qti.qualcomm.com>
Signed-off-by: Conner Huff <chuff@codeaurora.org>