drivers: rmnet_shs: Fix TCP OOO switching
Currently TCP flows are not paused when switching. This
causes more OOO packets than nessesary. The congestion
window is halved because of these OOO packets
and the flow which is now on a perf core has a lower
load and our throughput goes down needlessly.
Changes are to tweak TCP logic to pause flows and to
minimize OOO.
Instant rate switching also needs tweaking in order for
TCP to be switched between cores rapidly.

CRs-Fixed: 2383805
Change-Id: I6c194b2ef76ae43311a96546b832e5a327544ffb
Acked-by: Raul Martinez <mraul@qti.qualcomm.com>
Signed-off-by: Conner Huff <chuff@codeaurora.org>