drivers: rmnet: perf: fix callback registration
rmnet_perf was trying to register for DL marker before
the port variable that rmnet driver maintains is allocated.
This means that when dl marker callbacks happened, perf
module was not getting flushed. Now we register once an
rmnet_data netdevice is registered instead of based on
the physical net device event

Also, cleans up some of the initialization code so that
failures of allocations and registries are more obvious.

Finally, add ps mode callback so that perf driver can
flush more aggresively when ps mode is not engaged.
Currently this feature is gated by a debug flag. A
couple dependencies need to be resolved before this can
be enabled by default.

Change-Id: I5969070b7f87929dd1ded1eee2a58e2728beab22
Signed-off-by: Conner Huff <chuff@codeaurora.org>