drivers: rmnet_perf: Avoid stamping hash information on fragments
Stamping the hash information on fragmented packets leads to reassembly
problems later on due the head fragment and the subsequent ones having
different hash values. As such, we should avoid calculating this hash and
setting the SKB fields when encountering packet fragments.

Additionally, we only can avoid this calculation for any packet that is
not TCP or UDP, as we will never need to use it for flow node insertion.

Change-Id: I84a30b3a3d04dfe031a36e69d50cdfa0871be774
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>