drivers: rmnet_shs: Flow limit, WQ prio/bps tweak
Previously a flow node would be allocated per new flow seen by
driver. This leaves the possibility that if many ports are
probed in a short window that would cause memory to be allocated
for each packet. This could be exploited to cause low memory
errors until the flows are cleaned up.

We can mitigate this by setting a limit to the number of active flows
that we actively manage before we start ignoring new flows.

During Instant rate switching a whole core as loaded and moves
any flow that lands on that core to gold. This marking duration
is being reduced from 10 wq ticks to 3.

Driver was incorrectly doubling bps causing sysfs and logs to be
inaccurate. This fixes that so bps is counted correctly.

CRs-Fixed: 2406711
Change-Id: I4a653e07e823884e957cb383430cf94c07c3d3f4
Acked-by: Raul Martinez <mraul@qti.qualcomm.com>
Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>