drivers: rmnet_perf: Flush on any TCP option change
Instead of looking only for specific TCP options, compare the entire
TCP options segment of the header, and flush if there are any changes.

This prevents options that we don't specifically check for to avoid being
coalesced over multiple packets. This allows the MPTCP protocol to pass
through rmnet_perf properly, with each packet containing the original
MPTCP options as they arrive

Change-Id: Iec4ca77884fd5545225e5468f780723deffe99b6
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>