drivers: rmnet: perf: Determine lengths based on IP
Currently we are determining lengths of packets off
of map header lengths. This means that if NW or
modem sends us some extra garbage and we insert
that packet into flow nodes, later on we will
stamp a new length on the packet based on
incorrect map determined lengths. By switching to
IP based lengths, we are no longer prone to such
problems.

Change-Id: Ifdb7a906016cbc5bf74088fd736f1f22a7f2713a
Signed-off-by: Conner Huff <chuff@codeaurora.org>